code/management/python/requirements.txt
python-dotenv
sh
PyYAML